chore(python): Refactor connectors
Migrating to httpx and the latest changes to provisioning made it clear that the current architecture isn't the best one.

The point of the connectors was to match what I was seeing in the engine for connecting via http, unix, dial stdio, bin.

Current dispatch by `DAGGER_HOST` env var is really only used in provisioning. Client connection types (http vs unix socket) can be configured via httpx mounts.

This change decouples engine provisioning from establishing a client session. `dagger.Connection` combines both, but they can be used separately if needed.
