+++ sdk/python/src/dagger/.gitattributes
@@ -1,0 +1,1 @@
+/_engine.py linguist-generated=true
