+++ sdk/python/src/dagger/__init__.py
@@ -1,4 +1,5 @@
 from .api.gen import *
 from .api.gen_sync import Client as SyncClient
+from .config import Config as Config
 from .connection import Connection as Connection
-from .connectors import Config as Config
+from .exceptions import DaggerError as Error
