+++ sdk/python/src/dagger/api/base.py
@@ -13,7 +13,7 @@ from cattrs.preconf.json import make_converter
 from gql.client import AsyncClientSession, SyncClientSession
 from gql.dsl import DSLField, DSLQuery, DSLSchema, DSLSelectable, DSLType, dsl_gql
 
-from dagger.exceptions import DaggerException
+from dagger.exceptions import DaggerError
 
 _T = TypeVar("_T")
 
@@ -184,7 +184,7 @@ class Root(Type):
         return self._ctx.schema._schema
 
 
-class ClientError(DaggerException):
+class ClientError(DaggerError):
     """Base class for client errors."""
