--- sdk/python/src/dagger/connectors/.gitattributes
@@ -1,1 +1,0 @@
-/engine_version.py linguist-generated=true
