--- sdk/python/src/dagger/connectors/__init__.py
@@ -1,9 +1,0 @@
-from . import bin, docker, http  # noqa
-from .base import Config, Connector, get_connector, register_connector
-
-__all__ = [
-    "Config",
-    "Connector",
-    "get_connector",
-    "register_connector",
-]
