+++ sdk/python/src/dagger/engine/__init__.py
@@ -1,0 +1,7 @@
+from . import bin, docker  # noqa
+from .base import ProvisionError, get_engine
+
+__all__ = [
+    "get_engine",
+    "ProvisionError",
+]
