+++ sdk/python/src/dagger/exceptions.py
@@ -1,2 +1,2 @@
-class DaggerException(Exception):
+class DaggerError(Exception):
     """Base exception for all Dagger exceptions."""
