--- sdk/python/tests/connectors/test_base.py
@@ -1,6 +1,0 @@
-import dagger
-
-
-def test_timeout():
-    conn = dagger.Connection(dagger.Config(timeout=20))
-    assert conn.connector.cfg.timeout == 20
